macOS/clang fixes
The license.txt change is a line endings one. The .gitattributes tells
git to convert the LF stored in the repo to CRLF on check-in and
check-out. However, the file was stored in the repo *as CRLF*, so on
platforms other than Windows git would think the file had always been
modified, as it was comparing the CRLF internal stored version to the
CRLF-converted-to-LF checked-out version. The fix is to re-commit the
file with LF, which this commit does. Because of the .gitattributes, it
will however be checked-out as CRLF on all platforms as desired.

+++ hibikase/KaraokeData/SoramimiSong.cpp
@@ -147,7 +147,7 @@ void SoramimiLine::Serialize(const QVector<Syllable*>& syllables)
     m_raw_content += m_prefix;
 
     Centiseconds previous_time = Centiseconds::min();
-    size_t last_character_of_previous_text = 0;
+    int last_character_of_previous_text = 0;
     for (const Syllable* syllable : syllables)
     {
         Centiseconds start = syllable->GetStart();
